feat: create page for render component admin-service-add
