+++ src/components/admin/adminsidebar.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 import { useRouter } from "next/router";
 import { useState } from "react";
 
-export default function AdminSidebar() {
+export function AdminSidebar() {
   const [activeButton, setActiveButton] = useState("");
   const router = useRouter();
