+++ src/components/ui/IconArrowBack.tsx
@@ -1,0 +1,19 @@
+export function IconArrowBack() {
+    return (
+      <svg
+        width="40"
+        height="40"
+        viewBox="0 0 40 40"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M25.0007 31.6667L13.334 20L25.0007 8.33337"
+          stroke="#646C80"
+          strokeWidth="3"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+        />
+      </svg>
+    );
+  }
