+++ src/pages/adminservice/edit/[id].tsx
@@ -1,13 +1,10 @@
-import Adminsidebar from "@/components/admin/adminsidebar";
+import { AdminSidebar } from "@/components/admin/adminsidebar";
 import { useEffect, useState } from "react";
-import { useServices } from "@/components/ServicesContext";
 import { useRouter } from "next/router";
-import { FormEvent } from "react";
 import axios from "axios";
 import IconPicture from "@/components/ui/IconPicture";
 import IconDrag from "@/components/ui/IconDragAddAdmin";
 import IconPlusDefaultColor from "@/components/ui/IconPluseDefaultColor";
-import { stringify } from "querystring";
 import {
   Select,
   SelectContent,
@@ -17,6 +14,9 @@ import {
 } from "@/components/ui/select";
 import IconWarning from "@/components/ui/Iconwarning";
 import IconX from "@/components/ui/IconX";
+import IconCheck from "@/components/ui/IconCheck";
+import { IconArrowBack } from "@/components/ui/IconArrowBack";
+import { AdminSubmitPopUp } from "@/components/admin/adminSubmitPopUp";
 
 interface SubService {
   description: string;
@@ -24,30 +24,29 @@ interface SubService {
   unit_price: number;
 }
 
-export default function AdminNavbar() {
+export default function AdminEdit() {
+  // this is data input subservice for recive value form service index to request
   const [inputSubservice, setInputSubservice] = useState<SubService[]>([]);
-  // console.log(
-  //   "check subservice id and value of subservice when refresh window",
-  //   inputSubservice
-  // );
-  // const modifyInputSubservice = () => {
-  //   const updatedSubservices = inputSubservice.map((subservice) => ({
-  //     ...subservice,
-  //     unit_price: subservice.unit_price ?? 0, // กำหนดค่า unit_price เป็น 0 ถ้าเป็น null
-  //   }));
-  //   setInputSubservice(updatedSubservices);
-  // };
+  // console.log("check subservice data type before sent", inputSubservice);
+
+  // title service name for sent request
   const [inputTitle, setInputTitle] = useState<string>("");
+
+  // category id for sent request
   const [inputCat, setInputCat] = useState<number | undefined>();
   // console.log("this test for receive value inputCat", inputCat);
+
+  // file image to form.append for sent request
   const [inputImage, setInputImage] = useState<File>();
   // console.log("check image when update", inputImage);
+
+  // at the navbar for show name of service
   const [nameTopic, setNameTopic] = useState<String>("loading");
-  const [URLimage, setURLimage] = useState<String>();
-  // const [changeImage,setChangeImage] = useState<Boolean>(false)
-  const [showPopUpSubmit, setShowPopUpSubmit] = useState<Boolean>(false);
+
+  const [showPopUpSubmit, setShowPopUpSubmit] = useState<boolean>(false);
   const [showPopUpDeleteImg, setShowPopUpDeleteImg] = useState<Boolean>(false);
 
+  //ใช้สำหรับรับค่า id จาก endpoint params
   const router = useRouter();
   const { id } = router.query;
 
@@ -55,52 +54,39 @@ export default function AdminNavbar() {
     e.preventDefault();
 
     const formData = new FormData();
-    // formData.append("service_id", id: Number)
-
-    const serviceId = Array.isArray(id) ? id[0] : id; // เลือกค่าตัวแรกจากอาร์เรย์ ถ้า id เป็นอาร์เรย์
+    const serviceId = id;
 
     if (serviceId) {
-      formData.append("service_id", serviceId.toString()); // แปลงเป็น string ถ้าไม่ใช่
+      formData.append("service_id", serviceId.toString());
     }
-
     formData.append("title", inputTitle);
     if (inputCat) {
       formData.append("category_id", inputCat.toString());
     }
-
     if (inputImage) {
       formData.append("image", inputImage);
     }
-
+    // subservice is array of object must turn to be json to sent
     formData.append("subservices", JSON.stringify(inputSubservice));
 
-    console.log("FormData contents:");
-    let formDataObject: { [key: string]: any } = {};
-    for (let [key, value] of formData.entries()) {
-      formDataObject[key] = value;
-    }
-
-    console.log(
-      "test data for sent request by form data.append",
-      formDataObject
-    );
+    // loop in form for check data to sent
+    // let formDataObject: { [key: string]: any } = {};
+    // for (let [key, value] of formData.entries()) {
+    //   formDataObject[key] = value;
+    // }
+    // console.log("FormData contents: ", formDataObject);
 
     // Commented out API call
     try {
       await axios.put(`/api/admin/management/edit/${id}`, formData, {
         headers: { "Content-Type": "multipart/form-data" },
       });
-      router.push("/adminservice");
-      console.log("fromdata2", formData);
+      // console.log("fromdata2", formData);
+      setShowPopUpSubmit(true);
     } catch (error) {
       console.log(error);
     }
   };
-
-  // // ตัวอย่างการเรียกใช้ modifyInputSubservice
-  // useEffect(() => {
-  //   modifyInputSubservice(); // เรียกใช้ฟังก์ชัน
-  // }, [handleSubmit]); // เรียกใช้เมื่อ inputSubservice เปลี่ยนแปลง
 
   return (
     <>
@@ -114,8 +100,9 @@ export default function AdminNavbar() {
         }}
       >
         <div className="flex flex-row w-full">
+          {/* admin sidebar */}
           <div>
-            <Adminsidebar />
+            <AdminSidebar />
           </div>
           <div className="w-full flex flex-col">
             {/* navbar for admin page */}
@@ -124,13 +111,13 @@ export default function AdminNavbar() {
                 className="flex flex-row items-center cursor-pointer"
                 onClick={() => router.push("/adminservice")}
               >
-                <ArrowBack />
-                <div className="flex flex-col">
+                <IconArrowBack />
+                <nav className="flex flex-col">
                   <div className="text-xs">บริการ</div>
                   <div className="text-xl font-medium">{nameTopic}</div>
-                </div>
+                </nav>
               </div>
-              <div className="h-full flex flex-row items-center gap-6 relative">
+              <nav className="h-full flex flex-row items-center gap-6 relative">
                 <button
                   className="bg-white text-defaultColor text-base h-full px-7 flex items-center gap-3 rounded-lg w-32 text-center justify-center border border-defaultColor"
                   onClick={() => router.push("/adminservice")}
@@ -144,14 +131,14 @@ export default function AdminNavbar() {
                 >
                   ยืนยัน
                 </button>
-              </div>
+              </nav>
             </div>
+            {/* admin index */}
             <AdminserviceIndex
               setInputSubservice={setInputSubservice}
               setInputTitle={setInputTitle}
               setInputCat={setInputCat}
               SetInputimage={setInputImage}
-              setURLimage={setURLimage}
               setNameTopic={setNameTopic}
               setShowPopUpDeleteImg={setShowPopUpDeleteImg}
               showPopUpDeleteImg={showPopUpDeleteImg}
@@ -160,6 +147,15 @@ export default function AdminNavbar() {
           </div>
         </div>
       </form>
+
+      <AdminSubmitPopUp
+        setShowPopUpSubmit={setShowPopUpSubmit}
+        showPopUpSubmit={showPopUpSubmit}
+        message="แก้ไขรายการสำเร็จ"
+        subMessage="กรุณากดยืนยันเพื่อกลับสู่หน้าหลัก ?"
+        confirmationText="ยืนยัน"
+        redirectPath="/adminservice"
+      />
     </>
   );
 }
@@ -169,7 +165,6 @@ export const AdminserviceIndex = ({
   setInputTitle,
   setInputCat,
   SetInputimage,
-  setURLimage,
   setNameTopic,
   setShowPopUpDeleteImg,
   showPopUpDeleteImg,
@@ -178,28 +173,18 @@ export const AdminserviceIndex = ({
   const router = useRouter();
   const { id } = router.query;
 
+  // state for store subservice
   const [subservices, setSubservices] = useState<any[]>([]);
-  
-   useEffect(() => {
-    const updatedSubservices = subservices.map((subService) => ({
-      ...subService,
-      unit_price: subService.unit_price === "" ? "0" : subService.unit_price,
-    }));
-  
-    setInputSubservice(updatedSubservices);
-  }, [subservices]);
 
-  //  useEffect(() => {
-  //   setInputSubservice(subservices);
-  // }, [subservices]);
-
-  const [dataParams, setDataParams] = useState();
-  const [serviceNameData, setServiceNameData] = useState<String>("");
-  const [serviceCategoryData, setServiceCategoryData] = useState<String>();
+  // store all of data categories for show
   const [fetchDataCategories, setFetchDataCategories] = useState<any>([]);
+
+  // for recive data response to show data detail before sent request
+  const [serviceCategoryData, setServiceCategoryData] = useState<String>();
   const [createAt, setCreateAt] = useState<string>(new Date().toISOString());
   const [updateAt, setUpdateAt] = useState<string>(new Date().toISOString());
 
+  // add new row for subservice
   const addSubService = () => {
     setSubservices((prevSubservices) => [
       ...prevSubservices,
@@ -207,11 +192,13 @@ export const AdminserviceIndex = ({
     ]);
   };
 
+  // function for filter idx that not match with index
   const deleteSubservice = (index: number) => {
     const updatedSubservices = subservices.filter((_, idx) => idx !== index);
     setSubservices(updatedSubservices);
   };
 
+  // ค่าที่ได้รับการเปลี่ยนแปลงส่งเข้าไปเก็บข้องใน setSubservices
   const updateSubservice = (
     index: number,
     field: string,
@@ -226,6 +213,16 @@ export const AdminserviceIndex = ({
       return updatedSubservices;
     });
   };
+
+  // แปลงข้อมูลที่ได้รับมาจาก subservice ก่อนที่จะส่งให้ setInputSubservice โดยที่ unit_price ห้ามเป็น null
+  useEffect(() => {
+    const updatedSubservices = subservices.map((subService) => ({
+      ...subService,
+      unit_price: subService.unit_price === "" ? "0" : subService.unit_price,
+    }));
+    // console.log("updata subservice", updatedSubservices);
+    setInputSubservice(updatedSubservices);
+  }, [subservices]);
 
   const handleInputTitle = (event: React.ChangeEvent<HTMLInputElement>) => {
     // setServiceNameData(event.target.value);
@@ -285,8 +282,7 @@ export const AdminserviceIndex = ({
         `/api/admin/management/selectedit/${id}`
       );
       console.log("test response fetching data", response.data);
-      setDataParams(response.data);
-      setServiceNameData(response.data.service_name);
+
       setNameTopic(response.data.service_name);
       setServiceCategoryData(response.data.categories.category);
       // setSubservices(response.data.sub_services);
@@ -299,7 +295,7 @@ export const AdminserviceIndex = ({
           }) => ({
             ...subService,
             unit_price:
-              subService.unit_price === 0 ? null : subService.unit_price,
+              subService.unit_price === 0 ? "" : subService.unit_price,
           })
         )
       );
@@ -324,7 +320,7 @@ export const AdminserviceIndex = ({
 
   const fetchCategories = async () => {
     try {
-      const response = await axios.get(`/api/admin/management/getCategories`);
+      const response = await axios.get(`/api/admin/management/get-categories`);
       setFetchDataCategories(response.data);
     } catch (error) {
       console.log(error);
@@ -496,20 +492,6 @@ export const AdminserviceIndex = ({
               <div className="flex flex-col justify-between w-full gap-5 ">
                 <div className="flex gap-2">
                   <div>
-                    {new Date(updateAt).toLocaleDateString("en-US", {
-                      year: "numeric",
-                      month: "2-digit",
-                      day: "2-digit",
-                    })}
-                  </div>
-                  {new Date(updateAt).toLocaleTimeString("en-US", {
-                    hour: "2-digit",
-                    minute: "2-digit",
-                    hour12: true,
-                  })}
-                </div>
-                <div className="flex gap-2 ">
-                  <div>
                     {new Date(createAt).toLocaleDateString("en-US", {
                       year: "numeric",
                       month: "2-digit",
@@ -517,6 +499,20 @@ export const AdminserviceIndex = ({
                     })}
                   </div>
                   {new Date(createAt).toLocaleTimeString("en-US", {
+                    hour: "2-digit",
+                    minute: "2-digit",
+                    hour12: true,
+                  })}
+                </div>
+                <div className="flex gap-2 ">
+                  <div>
+                    {new Date(updateAt).toLocaleDateString("en-US", {
+                      year: "numeric",
+                      month: "2-digit",
+                      day: "2-digit",
+                    })}
+                  </div>
+                  {new Date(updateAt).toLocaleTimeString("en-US", {
                     hour: "2-digit",
                     minute: "2-digit",
                     hour12: true,
@@ -582,6 +578,7 @@ export function AddSubService({
   };
 
   deleteSubservice: (index: number) => void;
+
   updateSubservice: (
     index: number,
     field: "description" | "unit" | "unit_price",
@@ -653,25 +650,5 @@ export function AddSubService({
         </h1>
       </div>
     </>
-  );
-}
-
-export function ArrowBack() {
-  return (
-    <svg
-      width="40"
-      height="40"
-      viewBox="0 0 40 40"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M25.0007 31.6667L13.334 20L25.0007 8.33337"
-        stroke="#646C80"
-        strokeWidth="3"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
   );
 }
